+++ packages/cli/src/shims/fs.node.js
@@ -7,6 +7,7 @@ export {
   readdirSync,
   readFileSync,
   rmSync,
+  statfsSync,
   statSync,
   unlinkSync,
   writeFileSync,